Porting over changes from rethinkdb helm chart.
- The path to the probe has changed from `/probe` to `/rethinkdb-probe`.
- Add RETHINKDB_PASSWORD environment variable to set `--initial-password`.
- Remove `ready-probe.sh` because I don't think it's being used.
- A bit of re-organizing of the files.
- Update version of Docker Tag to 2.3.5-v2.
