--- Dockerfile.build
@@ -1,10 +1,0 @@
-FROM golang:1.6
-MAINTAINER Chris Dornsife <chris@applariat.com>
-
-# Build container to have a consistent go build environment
-
-COPY . /go/src/probe
-WORKDIR /go/src/probe
-
-RUN go get ./... \
-    && CGO_ENABLED=0 GOOS=linux go build -a -ldflags '-w -s' -o /target/probe .
